// client/src/types.ts
export type UserRole = 'ADMIN' | 'MEMBER' | 'SUPER_ADMIN' | 'LIBRARY_ADMIN' | 'LIBRARIAN';
// ...
export interface User {
  id: string;
  fullName: string;
  email: string;
  username?: string;
  mobile: string;
  address: string;
  role: UserRole;
  passwordHash: string;
  memberId: string;
  createdAt: string;
  status: 'active' | 'suspended' | 'pending';
  avatarUrl?: string;
  mustChangePassword?: boolean;
  isEmailVerified?: boolean;
  twoFactorEnabled?: boolean;
}